api_test/learn_qa/test_api_upd/requirements.txt
allallure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
requests==2.31.0
setuptools==69.0.2
urllib3==2.1.0
